hmac: fix output
HMAC was fundamentally broken, the major cause was messages we're being
digested first and then passed to the HMAC routine. This caused external
code to not be able to verify HMAC signatures. Fix this by using the
TPM2_HMAC command.

Further enhancements to support arbitrary size buffers will need to use
StartHMAC session and chunk the data over.

Signed-off-by: William Roberts <william.c.roberts@intel.com>