tpm: fix suspicious sizeof
The sizeof on dest in str_padded_copy requires the src to be a fixed
buffer. Since the vendor list is an array of const char pointers,
sizeof on the value is the char pointer not an array of chars. Fix this
by using the _ version where we can control the inputs and use
strlen here since the buffer is gaurenteed to be NUL term over sizeof.

Signed-off-by: William Roberts <william.c.roberts@intel.com>