tools: fix try/finally scope around os.write
Doing:

    try:
        privfd, tertiarypriv = mkstemp(prefix='', suffix='.priv', dir=d)
        # ...
    finally:
        os.close(privfd)

can lead to issues when the call to `mkstemp` raises an exception,
as `privfd` does not exist yet. Fix this issue by moving `mkstemp` out
of the `try`/`finally` block.

When writing both the private and public files, in class `LinkCommand`,
this change leads to intricated (and hard-to-understand) `try`/`finally`
blocks. Reduce the complexity of the code by writing the files one after
another.

Finally, introduce a `try`/`finally` around an `os.write` in function
`importkey`, as it makes sense if the operating system raises an
exception while writing the file (for example when the storage is full).

Signed-off-by: Nicolas Iooss <nicolas.iooss@ledger.fr>