test/openssl.sh: check that importing a certificate does not change it
Create a new certificate which is signed by an existing one, using
extensions which are encoded as booleans. Then use `openssl verify` to
check that the signature of the certificate is still valid, when they
are imported in tpm2-pkcs11's database.

When `tpm2_ptool addcert` re-encode the certificates, using for example:

    bercert = pyasn1.codec.ber.encoder.encode(cert)

in function `pemcert_to_attrs`, this new test fails with:

    + openssl verify -CAfile ca.pem cert.pem
    CN = my sub key
    error 7 at 0 depth lookup: certificate signature failure
    140336169449280:error:04091068:rsa routines:int_rsa_verify:bad signature:crypto/rsa/rsa_sign.c:228:
    140336169449280:error:0D0C5006:asn1 encoding routines:ASN1_item_verify:EVP lib:crypto/asn1/a_verify.c:179:
    error cert.pem: verification failed
    ++ onerror
    ++ echo 'openssl verify -CAfile ca.pem cert.pem on line 149 failed: 2'
    openssl verify -CAfile ca.pem cert.pem on line 149 failed: 2
    ++ exit 1

Signed-off-by: Nicolas Iooss <nicolas.iooss@ledger.fr>