Fix C_Encrypt() and C_Decrypt() when using CKM_AES_CBC_PAD
When using pkcs11 it is a common pattern to encrypt data like this:

  CK_ULONG_PTR out_len;
  // find the output length
  C_Encrypt(h, in, in_len, NULL, out_len);
  // allocate buffer with out_len bytes and do the encryption

The actual encryption is done by either a single call to C_Encrypt() or
multiple calls to C_EncryptUpdate(). In both cases we use common
functions which keep internal state needed for CKM_AES_CBC_PAD.
The problem is that C_Encrypt() modifies the state even in the case when
it is used for finding the output length. This causes a subsequent call
to C_Encrypt() for the actual encryption to fail.

The same problem exist for C_Decrypt().

This patch fixes the above problems by clearing the internal state at
the beginning of C_Encrypt/C_Decrypt.

Fixes: #730
Signed-off-by: Radoslav Gerganov <rgerganov@vmware.com>