ci: fix perms between clone and docker
Docker runs in a root user container and the source code checkout is
done with some other UID. This causes git to complain its not a safe
directory so we need to mark it safe.

See the below link for details:
  - https://github.blog/2022-04-12-git-security-vulnerability-announced/

This issue is not relevant for our disposable containers so we can
safely ignore this.

Signed-off-by: William Roberts <william.c.roberts@intel.com>