db: fix upgrade backup
During a failed upgrade, the original database was deleted and replaced
with the upgraded one, making it impossible to revert the failed
upgrade.

This commit fixes this problem by keeping the old version of the
database as a separate file for upgrades that finished successfully and
keeping the original database for those that did not.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>